Move holiday object to runtime data in workday
